+++ IBAnimatable.playground/Pages/Predefined Animations.xcplaygroundpage/Contents.swift
@@ -26,11 +26,17 @@ view.borderWidth = 2
 view.borderColor = UIColor.purpleColor()
 view.maskType = "Circle"
 
+// For moveTo or moveBy animation
+view.x = -100
+view.y = 200
 
 //: Animations, all supported predefined animations can be found in `enum AnimationType`  
 
 // Uncomment one line to play the animation
 
+//view.moveX()
+//view.moveY()
+//view.moveXY()
 //view.slideInLeft()
 //view.slideInRight()
 //view.slideInDown()
@@ -70,7 +76,7 @@ view.maskType = "Circle"
 //view.zoomIn()
 //view.zoomOut()
 //view.shake()
-view.pop()
+//view.pop()
 //view.flipX()
 //view.flipY()
 //view.morph()
@@ -80,5 +86,7 @@ view.pop()
 //view.swing()
 //view.rotate()
 //view.rotate(clockwise: false)
+//view.moveTo()
+view.moveBy()
 
 //: [Next](@next)