+++ IBAnimatable/AnimatableLabel.swift
@@ -16,6 +16,8 @@ import UIKit
   @IBInspectable public var velocity: CGFloat = CGFloat.NaN
   @IBInspectable public var force: CGFloat = CGFloat.NaN
   @IBInspectable public var repeatCount: Float = Float.NaN
+  @IBInspectable public var x: CGFloat = CGFloat.NaN
+  @IBInspectable public var y: CGFloat = CGFloat.NaN
   
   // MARK: - RotationDesignable
   @IBInspectable public var rotate: CGFloat = CGFloat.NaN {